Slight modifications before pushing
