+++ Cargo.toml
@@ -22,7 +22,7 @@ r2g_mlua = { version = "0.10.3", features = ["luau", "luau-jit", "macros", "asyn
 parking_lot = { version = "0.12.3", features = [] }
 rustversion = "1.0.19"
 rustversion-detect = "0.1.3"
-roblox-to-godot-project-derive = { path = "./roblox-to-godot-project-derive" }
+rblx-godot-derive = { path = "./rblx-godot-derive" }
 
 [workspace]
