--- roblox-to-godot-project-derive/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "roblox-to-godot-project-derive"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-convert_case = "0.7.1"
-proc-macro2 = "1.0.93"
-quote = "1.0.38"
-syn = { version = "2.0.96", features = ["full", "visit"] }
-
-[lib]
-proc-macro = true
